+++ src/Console/CreateIndex.php
@@ -4,7 +4,9 @@ namespace Elodex\Console;
 
 use Illuminate\Console\Command;
 use Illuminate\Console\ConfirmableTrait;
+use Illuminate\Support\Str;
 use Elodex\IndexManager;
+use Elodex\Contracts\IndexedModel as IndexedModelContract;
 
 class CreateIndex extends Command
 {
@@ -19,6 +21,7 @@ class CreateIndex extends Command
                             {--I|index= : The name of the index to create}
                             {--S|shards= : Number of shards}
                             {--R|replicas= : Number of replicas}
+                            {--models= : Comma separated list of indexed model classes used for property mappings}
                             {--reset : Reset index if it already exists}
                             {--force : Force the operation to run when in production}';
 
@@ -59,6 +62,7 @@ class CreateIndex extends Command
         $indexName = $this->option('index') ?: $this->indexManager->getDefaultIndex();
         $reset = $this->option('reset') ?: false;
 
+        // Check if the specified index already exists.
         if ($this->indexManager->indicesExist([$indexName])) {
             if ($reset && $this->confirmToProceed()) {
                 $this->resetIndex($indexName);
@@ -69,6 +73,7 @@ class CreateIndex extends Command
             }
         }
 
+        // Build the settings array.
         $settings = [];
         $shards = $this->option('shards');
         if (! is_null($shards)) {
@@ -80,9 +85,71 @@ class CreateIndex extends Command
             $settings['number_of_replicas'] = $replicas;
         }
 
-        $this->createIndex($indexName, $settings);
+        $models = [];
+        if ($this->option('models')) {
+            $models = explode(',', trim($this->option('models')));
+
+            if (($models = $this->parseModelsOption($models)) === false) {
+                return 1;
+            }
+        }
+
+        $this->createIndex($indexName, $settings, $models);
 
         $this->info("Index '{$indexName}' successfully created.");
+    }
+
+    /**
+     * Checks if the specified models are valid indexed model classes and builds
+     * an array of fully qualified class names.
+     *
+     * @param  array $models
+     * @return bool|array
+     */
+    protected function parseModelsOption(array $models)
+    {
+        $parsed = [];
+
+        foreach ($models as $model) {
+            $class = $this->parseModelName(trim($model));
+
+            if (! class_exists($class)) {
+                $this->error("Model class '{$class}' does not exist!");
+
+                return false;
+            }
+
+            if (! ((new $class) instanceof IndexedModelContract)) {
+                $this->error("Model class '{$class}' does not implement the '".IndexedModelContract::class."' interface!");
+
+                return false;
+            }
+
+            $parsed[] = $class;
+        }
+
+        return $parsed;
+    }
+
+    /**
+     * Parse the model name and format according to the root namespace.
+     *
+     * @param  string  $name
+     * @return string
+     */
+    protected function parseModelName($name)
+    {
+        $rootNamespace = $this->laravel->getNamespace();
+
+        if (Str::startsWith($name, $rootNamespace)) {
+            return $name;
+        }
+
+        if (Str::contains($name, '/')) {
+            $name = str_replace('/', '\\', $name);
+        }
+
+        return $this->parseModelName(trim($rootNamespace, '\\').'\\'.$name);
     }
 
     /**
@@ -101,10 +168,38 @@ class CreateIndex extends Command
      *
      * @param  string $indexName
      * @param  array $settings
+     * @param  array $models
      * @return void
      */
-    protected function createIndex($indexName, $settings)
+    protected function createIndex($indexName, $settings, array $models = [])
     {
-        $this->indexManager->createIndex($indexName, $settings);
+        $mappings = null;
+        if (! empty($models)) {
+            $mappings = $this->getPropertyMappings($models);
+        }
+
+        $this->indexManager->createIndex($indexName, $settings, $mappings);
+    }
+
+    /**
+     * Get the property mappings for the models.
+     *
+     * @param  array $models
+     * @return array
+     */
+    protected function getPropertyMappings(array $models = [])
+    {
+        $mappings = [];
+
+        // Add the default property mappings for the model classes.
+        foreach ($models as $class) {
+            $model = new $class;
+            $mappings[$model->getIndexTypeName()] = [
+                '_source' => ['enabled' => true],
+                'properties' => $model->getIndexMappingProperties()
+            ];
+        }
+
+        return $mappings;
     }
 }