+++ src/IndexServiceProvider.php
@@ -9,6 +9,7 @@ use Elodex\Console\DeleteIndex;
 use Elodex\Console\CreateIndex;
 use Elodex\Console\GetMappings;
 use Elodex\Console\GetStats;
+use Elodex\Console\GetSettings;
 use Elodex\Console\Upgrade;
 use Elodex\Console\Analyze;
 use Elodex\Console\MakeSyncHandler;
@@ -78,9 +79,9 @@ class IndexServiceProvider extends ServiceProvider
     {
         $this->app->singleton(IndexManager::class, function ($app) {
             $client = $app[ElasticsearchClientManager::class];
-            $defaultIndex = $app['config']->get('elodex.default_index', 'default');
+            $config = $app['config']->get('elodex');
 
-            return new IndexManager($client, $defaultIndex);
+            return new IndexManager($client, $config);
         });
 
         $this->app->alias(IndexManager::class, 'elodex.index');
@@ -130,6 +131,9 @@ class IndexServiceProvider extends ServiceProvider
         $this->app->singleton(GetStats::class, function () use ($indexManager) {
             return new GetStats($indexManager);
         });
+        $this->app->singleton(GetSettings::class, function () use ($indexManager) {
+            return new GetSettings($indexManager);
+        });
         $this->app->singleton(Upgrade::class, function () use ($indexManager) {
             return new Upgrade($indexManager);
         });
@@ -147,6 +151,7 @@ class IndexServiceProvider extends ServiceProvider
             DeleteIndex::class,
             GetMappings::class,
             GetStats::class,
+            GetSettings::class,
             Upgrade::class,
             Analyze::class,
             Seed::class,